Fix Entry in CheckVTableHasNoDuplicates().
The Entry constructor initialized Entry.name_len in a call
to another function as part of the Entry.name initialization
but given the order of members, the Entry.name_len was later
overwritten with 0. When the change
    https://android-review.googlesource.com/1535207
replaced strcmp() with memcmp(), relying on Entry.name_len
being properly initialized, the equality comparison was
broken. We fix that by initializing Entry.name_len before
the Entry.name and also adjust it for non-ASCII characters.

Test: m dump-oat  # No unexpected logs.
Bug: 175869411
Change-Id: I08f3425b4dac25adddc4f720c110d85533b99d8b
Signed-off-by: Akash Srivastava <akashniki@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>